fixed print

+++ src/main/java/com/github/paolodenti/jsapp/core/command/base/SappResponse.java
@@ -72,13 +72,13 @@ public class SappResponse {
 	/**
 	 *  @return interpreted response data, 2 bytes are interpreted as hex-ascii bytes in order to build a byte
 	 */
-	public int getDataAsByte() {
+	public byte getDataAsByte() {
 
-		int result = 0;
+		byte result = 0;
 
 		for (int i = 0; i < 2; i++) {
 			if (i < data.length) {
-				result = result << 4;
+				result = (byte) (result << 4);
 				result += SappUtils.getByteFromHexAsciiCode(data[i]);
 			} else {
 				break; // premature end of data
@@ -110,19 +110,19 @@ public class SappResponse {
 	/**
 	 *  @return interpreted response data, 2 bytes couples are interpreted as hex-ascii bytes in order to build a byte array
 	 */
-	public int[] getDataAsByteArray() {
+	public byte[] getDataAsByteArray() {
 
 		if (data.length < 2) {
 			return null;
 		}
 
-		int[] resultArr = new int[data.length / 2];
+		byte[] resultArr = new byte[data.length / 2];
 
 		for (int i = 0; i < data.length; i += 2) {
-			int result = 0;
+			byte result = 0;
 			for (int j = 0; j < 2; j++) {
 				if (i + j < data.length) {
-					result = result << 4;
+					result = (byte) (result << 4);
 					result += SappUtils.getByteFromHexAsciiCode(data[i + j]);
 				} else {
 					break; // premature end of data
